move to lib/time
