+++ protocol/daemons/server/types/bridge/bridge_event_manager.go
@@ -6,8 +6,8 @@ import (
 	"time"
 
 	"github.com/cosmos/cosmos-sdk/telemetry"
-	"github.com/dydxprotocol/v4-chain/protocol/lib"
 	"github.com/dydxprotocol/v4-chain/protocol/lib/metrics"
+	libtime "github.com/dydxprotocol/v4-chain/protocol/lib/time"
 	"github.com/dydxprotocol/v4-chain/protocol/x/bridge/types"
 )
 
@@ -29,7 +29,7 @@ type BridgeEventManager struct {
 	recognizedEventInfo types.BridgeEventInfo
 
 	// Time provider than can mocked out if necessary
-	timeProvider lib.TimeProvider
+	timeProvider libtime.TimeProvider
 }
 
 // BridgeEventWithTime is a type that wraps BridgeEvent but also
@@ -41,7 +41,7 @@ type BridgeEventWithTime struct {
 
 // NewBridgeEventManager creates a new BridgeEventManager.
 func NewBridgeEventManager(
-	timeProvider lib.TimeProvider,
+	timeProvider libtime.TimeProvider,
 ) *BridgeEventManager {
 	return &BridgeEventManager{
 		events: make(map[uint32]BridgeEventWithTime),
